requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.2.0
pydantic==2.7.4
openai==1.35.7
scikit-learn==1.5.1
PyJWT==2.8.0
python-multipart==0.0.9
